src/mixmeasurebuddy_webapp/api/requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.6
dnspython==2.4.1
email-validator==2.0.0.post2
flask-mongoengine==1.0.0
flask-sqlalchemy==3.0.5
flask-swagger-generator==0.4
flask-swagger-ui==4.11.1
Flask-WTF==1.1.1
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mongoengine==0.27.0
packaging==23.1
pip==23.2.1
pydantic-core==2.4.0
pydantic==2.1.1
pymongo==4.4.1
python-dotenv==1.0.0
#setuptools==65.5.1
SQLAlchemy==2.0.19
typing-extensions==4.7.1
Werkzeug==2.3.6
wheel==0.38.4
WTForms==3.0.1
zipp==3.16.2
names>=0.3.0